requirements.txt
amqp==5.0.6
aniso8601==8.1.1
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
defusedxml==0.7.1
Django==3.2.2
django-celery-beat==2.2.0
django-filter==2.4.0
django-graphql-jwt==0.3.2
django-redis==4.12.1
django-taggit==1.4.0
django-timezone-field==4.1.2
graphene==3.0b7
graphene-django==3.0.0b7
graphene-file-upload==1.3.0
graphql-core==3.1.5
graphql-relay==3.1.0
gunicorn==20.1.0
idna==2.10
isodate==0.6.0
kombu==5.0.2
lxml==4.6.3
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.18
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zeep==4.0.0
psycopg2-binary==2.8.6
